feat: credit card validation added

+++ CreditCard.cs
@@ -2,12 +2,12 @@
 {
     public class CreditCard
     {
-        public int CreditCardNumber { get; set; }
+        public long CreditCardNumber { get; set; }
         public DateTime ExpirationDate { get; set; }
         public int CvvNumber { get; set; }
         public string CardholderName { get; set; }
 
-        public CreditCard(int creditcardnumber, DateTime expirationDate, int cvvNumber, string cardholdername)
+        public CreditCard(long creditcardnumber, DateTime expirationDate, int cvvNumber, string cardholdername)
         {
             CreditCardNumber = creditcardnumber;
             ExpirationDate = expirationDate;
